Cleaned up the serialization a bit.
If we're doing CSV serialization (which we always are) it makes sense
to handle that logic within the CSV serialization step.

The current logic flips the old process of "finding the maximum" before
"trimming flushed values" this helps b/c it means we typically output
fewer rows. Halving the output data where previously it was just a bunch
of "0.000" entries.