Update Spawning
Actually ran an analysis today that was large enough I could see
how spawning was happening. Come to find out, lazy iteration
means nothing will actually happen in parallel, as once an iterator
spawns the thread, it waits on it.

The code has churned a bit to remove crossbeam, and update the
thread spawning behavior.